Merge changes

+++ src/bf/mean.rs
@@ -2,7 +2,7 @@ use std::ops;
 use super::num;
 
 pub fn mean<T>(list: &[T]) -> T
-               where T: num::Num + ops::Add<T, Output=T> + From<usize> + Clone {
+               where T: num::Num + ops::Add<T, Output=T> + From<f64> + Clone {
     let sum: T = {
         // Perhaps replace with iterator? Though might be slower/inefficient
         let mut s_sum: T = T::zero();
@@ -11,7 +11,7 @@ pub fn mean<T>(list: &[T]) -> T
         }
         s_sum
     };
-    sum / T::from(list.len())
+    sum / T::from(list.len() as f64)
 }
 
 pub fn f_mean(list: &[f64]) -> f64 {
